fix: resolve VideoStat Int overflow for high view counts, UP space data retention, and Mobile BvPlayer lastPlayed sync

+++ app/shared/src/main/kotlin/dev/aaa1115910/bv/viewmodel/TagViewModel.kt
@@ -62,8 +62,8 @@ class TagViewModel : ViewModel() {
                         title = tagVideoItem.title,
                         cover = tagVideoItem.pic,
                         upName = tagVideoItem.owner.name,
-                        play = tagVideoItem.stat.view,
-                        danmaku = tagVideoItem.stat.danmaku,
+                        play = tagVideoItem.stat.view.toInt(),
+                        danmaku = tagVideoItem.stat.danmaku.toInt(),
                         time = tagVideoItem.duration.toLong()
                     )
                 )
